Remove use of radius parameter and transform scaling
The radius parameter was useful for the basic shapes like cylinder and sphere to scale it from unit size,
but with more fine control of the geometry through other parameters it is not necessary.

Also arbitrary scaling through the m_to_world transform can destroy assumptions in the geometry calculations so do not allow that,
require that scaling is an identity matrix.

Translation and translation through the m_to_world transform is still allowed and useful (but rotation may need a bit of fixing)